Fix undefined behaviour in implicit conversions
Cleans up errors in raster and topology and adds an
extra sanitizer build in Travis

Closes #4383
Closes https://github.com/postgis/postgis/pull/398

git-svn-id: http://svn.osgeo.org/postgis/trunk@17480 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/liblwgeom_topo.h
@@ -225,7 +225,7 @@ typedef struct LWT_BE_CALLBACKS_T {
 					   double dist,
 					   uint64_t *numelems,
 					   int fields,
-					   int limit);
+					   int64_t limit);
 
   /**
    * Insert nodes
@@ -285,7 +285,7 @@ typedef struct LWT_BE_CALLBACKS_T {
 					   double dist,
 					   uint64_t *numelems,
 					   int fields,
-					   int limit);
+					   int64_t limit);
 
   /**
    * Get next available edge identifier
